+++ benchmarks/Program.cs
@@ -1,11 +1,12 @@
 using System;
-using System.IO;
-using System.Linq;
 using System;
 using System.IO;
+using System.IO;
+using System.Linq;
 using System.Linq;
 
 using BenchmarkDotNet.Running;
+
 using Veggerby.Boards.Benchmarks;
 
 // Benchmark harness entry point.